chore(chat messages): move coroutine to its own file
