+++ ui/src/components/chat/compose/messages/coroutines.rs
@@ -1,0 +1,194 @@
+use common::{
+    state::{Action, State},
+    warp_runner::{RayGunCmd, WarpCmd},
+    WARP_CMD_CH,
+};
+
+use dioxus_core::Scoped;
+use dioxus_hooks::{to_owned, use_coroutine, Coroutine, UnboundedReceiver, UseRef, UseSharedState};
+use futures::StreamExt;
+
+use warp::raygun::{PinState, ReactionState};
+
+use super::{get_messagesProps, DownloadTracker, MessagesCommand, NewelyFetchedMessages};
+
+pub fn handle_warp_commands<'a>(
+    cx: &'a Scoped<'a, get_messagesProps>,
+    state: &UseSharedState<State>,
+    newely_fetched_messages: &UseRef<Option<NewelyFetchedMessages>>,
+    pending_downloads: &UseSharedState<DownloadTracker>,
+) -> Coroutine<MessagesCommand> {
+    let _ch = use_coroutine(cx, |mut rx: UnboundedReceiver<MessagesCommand>| {
+        to_owned![state, newely_fetched_messages, pending_downloads];
+        async move {
+            let warp_cmd_tx = WARP_CMD_CH.tx.clone();
+            while let Some(cmd) = rx.next().await {
+                match cmd {
+                    MessagesCommand::React((user, message, emoji)) => {
+                        let (tx, rx) = futures::channel::oneshot::channel();
+                        let reaction_state =
+                            match message.reactions().iter().find(|x| x.emoji() == emoji) {
+                                Some(reaction) if reaction.users().contains(&user) => {
+                                    ReactionState::Remove
+                                }
+                                _ => ReactionState::Add,
+                            };
+                        if let Err(e) = warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::React {
+                            conversation_id: message.conversation_id(),
+                            message_id: message.id(),
+                            reaction_state,
+                            emoji: emoji.clone(),
+                            rsp: tx,
+                        })) {
+                            log::error!("failed to send warp command: {}", e);
+                            continue;
+                        }
+
+                        let res = rx.await.expect("command canceled");
+                        match res {
+                            Ok(_) => state.write().mutate(Action::AddReaction(
+                                message.conversation_id(),
+                                message.id(),
+                                emoji,
+                            )),
+                            Err(e) => {
+                                log::error!("failed to add/remove reaction: {}", e);
+                            }
+                        }
+                    }
+                    MessagesCommand::DeleteMessage { conv_id, msg_id } => {
+                        let (tx, rx) = futures::channel::oneshot::channel();
+                        if let Err(e) =
+                            warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::DeleteMessage {
+                                conv_id,
+                                msg_id,
+                                rsp: tx,
+                            }))
+                        {
+                            log::error!("failed to send warp command: {}", e);
+                            continue;
+                        }
+
+                        let res = rx.await.expect("command canceled");
+                        if let Err(e) = res {
+                            log::error!("failed to delete message: {}", e);
+                        }
+                    }
+                    MessagesCommand::DownloadAttachment {
+                        conv_id,
+                        msg_id,
+                        file,
+                        file_path_to_download,
+                    } => {
+                        let (tx, rx) = futures::channel::oneshot::channel();
+                        if let Err(e) =
+                            warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::DownloadAttachment {
+                                conv_id,
+                                msg_id,
+                                file_name: file.name(),
+                                file_path_to_download,
+                                rsp: tx,
+                            }))
+                        {
+                            log::error!("failed to send warp command: {}", e);
+                            if let Some(conv) = pending_downloads.write().get_mut(&conv_id) {
+                                conv.remove(&file);
+                            }
+                            continue;
+                        }
+
+                        let res = rx.await.expect("command canceled");
+                        match res {
+                            Ok(mut stream) => {
+                                while let Some(p) = stream.next().await {
+                                    log::debug!("{p:?}");
+                                }
+                            }
+                            Err(e) => {
+                                log::error!("failed to download attachment: {}", e);
+                            }
+                        }
+                        if let Some(conv) = pending_downloads.write().get_mut(&conv_id) {
+                            conv.remove(&file);
+                        }
+                    }
+                    MessagesCommand::EditMessage {
+                        conv_id,
+                        msg_id,
+                        msg,
+                    } => {
+                        let (tx, rx) = futures::channel::oneshot::channel();
+                        if let Err(e) = warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::EditMessage {
+                            conv_id,
+                            msg_id,
+                            msg,
+                            rsp: tx,
+                        })) {
+                            log::error!("failed to send warp command: {}", e);
+                            continue;
+                        }
+
+                        let res = rx.await.expect("command canceled");
+                        if let Err(e) = res {
+                            log::error!("failed to edit message: {}", e);
+                        }
+                    }
+                    MessagesCommand::FetchMore {
+                        conv_id,
+                        to_fetch,
+                        current_len,
+                    } => {
+                        let (tx, rx) = futures::channel::oneshot::channel();
+                        if let Err(e) =
+                            warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::FetchMessages {
+                                conv_id,
+                                to_fetch,
+                                current_len,
+                                rsp: tx,
+                            }))
+                        {
+                            log::error!("failed to send warp command: {}", e);
+                            continue;
+                        }
+
+                        match rx.await.expect("command canceled") {
+                            Ok((messages, has_more)) => {
+                                newely_fetched_messages.set(Some(NewelyFetchedMessages {
+                                    conversation_id: conv_id,
+                                    messages,
+                                    has_more,
+                                }));
+                            }
+                            Err(e) => {
+                                log::error!("failed to fetch more message: {}", e);
+                            }
+                        }
+                    }
+                    MessagesCommand::Pin(msg) => {
+                        let (tx, rx) = futures::channel::oneshot::channel();
+                        let pinstate = if msg.pinned() {
+                            PinState::Unpin
+                        } else {
+                            PinState::Pin
+                        };
+                        if let Err(e) = warp_cmd_tx.send(WarpCmd::RayGun(RayGunCmd::Pin {
+                            conversation_id: msg.conversation_id(),
+                            message_id: msg.id(),
+                            pinstate,
+                            rsp: tx,
+                        })) {
+                            log::error!("failed to send warp command: {}", e);
+                            continue;
+                        }
+
+                        let res = rx.await.expect("command canceled");
+                        if let Err(e) = res {
+                            log::error!("failed to pin message: {}", e);
+                        }
+                    }
+                }
+            }
+        }
+    });
+    _ch.clone()
+}
